+++ Magic/CHANGELOG.md
@@ -38,6 +38,7 @@
  - Add `ignore_particles` mage option, allow turning off particles in the tweak menu
  - Reduce particles on flare, fountain and teleporter effects
  - Add `ignore_modified_entities` AreaOfEffect option, to ignore entities that have already been modified by this spell
+ - Add `game_mode` and `leave_game_mode` world option, for per-world game modes
 
 ## Changes